Used HookCache class instead of string.Join() to reduce garbage
