+++ Oxide.Core/Plugins/CSPlugin.cs
@@ -33,120 +33,6 @@ namespace Oxide.Core.Plugins
     /// </summary>
     public abstract class CSPlugin : Plugin
     {
-        public class HookMethod
-        {
-            public string Name;
-
-            public MethodInfo Method;
-
-            public ParameterInfo[] Parameters => Method.GetParameters();
-
-            public bool IsBaseHook => Name.StartsWith("base_");
-
-            public HookMethod(MethodInfo method)
-            {
-                Method = method;
-                Name = method.Name;
-
-                if (Parameters.Length > 0)
-                {
-                    Name += $"({string.Join(", ", Parameters.Select(x => x.ParameterType.ToString()).ToArray())})";
-                }
-            }
-            
-            public bool HasMatchingSignature(object[] args, out bool exact)
-            {
-                exact = true;
-
-                if (Parameters.Length == 0 && (args == null || args.Length == 0))
-                    return true;
-
-                for (var i = 0; i < args.Length; i++)
-                {
-                    if (args[i] == null)
-                    {
-                        if (CanAssignNull(Parameters[i].ParameterType))
-                        {
-                            continue;
-                        }
-
-                        return false;
-                    }
-
-                    if (exact)
-                    {
-                        if (args[i].GetType() != Parameters[i].ParameterType &&
-                            args[i].GetType().MakeByRefType() != Parameters[i].ParameterType &&
-                            !CanConvertNumber(args[i], Parameters[i].ParameterType))
-                        {
-                            exact = false;
-                        }
-                    }
-
-                    if (exact) continue;
-
-                    if (args[i].GetType() == Parameters[i].ParameterType ||
-                        args[i].GetType().MakeByRefType() == Parameters[i].ParameterType ||
-                        Parameters[i].ParameterType.FullName == "System.Object")
-                    {
-                        continue;
-                    }
-
-                    if (args[i].GetType().IsValueType)
-                    {
-                        if (!TypeDescriptor.GetConverter(Parameters[i].ParameterType).CanConvertFrom(args[i].GetType()) && !CanConvertNumber(args[i], Parameters[i].ParameterType))
-                        {
-                            return false;
-                        }
-                    }
-                    else
-                    {
-                        if (!Parameters[i].ParameterType.IsInstanceOfType(args[i]))
-                        {
-                            return false;
-                        }
-                    }
-                }
-
-                return true;
-            }
-            
-            private bool CanAssignNull(Type type)
-            {
-                if (!type.IsValueType)
-                {
-                    return true;
-                }
-
-                return Nullable.GetUnderlyingType(type) != null;
-            }
-
-            private bool IsNumber(object obj)
-            {
-                return obj != null && IsNumber(Nullable.GetUnderlyingType(obj.GetType()) ?? obj.GetType());
-            }
-
-            private bool IsNumber(Type type)
-            {
-                if (type.IsPrimitive)
-                {
-                    return type != typeof(bool) && type != typeof(char) && type != typeof(IntPtr) && type != typeof(UIntPtr);
-                }
-
-                return type == typeof(decimal);
-            }
-
-            private bool CanConvertNumber(object value, Type type)
-            {
-                if (!IsNumber(value) || !IsNumber(type))
-                {
-                    return false;
-                }
-
-                return TypeDescriptor.GetConverter(type).IsValid(value);
-            }
-        }
-
         /// <summary>
         /// Gets the library by the specified type or name
         /// </summary>
@@ -158,7 +44,7 @@ namespace Oxide.Core.Plugins
         protected Dictionary<string, List<HookMethod>> Hooks = new Dictionary<string, List<HookMethod>>();
 
         // All matched hooked methods
-        protected Dictionary<string, List<HookMethod>> HooksCache = new Dictionary<string, List<HookMethod>>();
+        protected HookCache HooksCache = new HookCache();
 
         /// <summary>
         /// Initializes a new instance of the CSPlugin class
@@ -230,7 +116,7 @@ namespace Oxide.Core.Plugins
         {
             object returnvalue = null;
             bool pooledArray = false;
-            
+
             // Call all hooks that match the signature
             foreach (var h in FindHooks(name, args))
             {
@@ -308,21 +194,15 @@ namespace Oxide.Core.Plugins
         protected List<HookMethod> FindHooks(string name, object[] args)
         {
             // Get the full name of the hook `name(argument type 1, argument type 2, ..., argument type x)`
-            var fullName = name;
-            if (args?.Length > 0)
-            {
-                fullName += $"({string.Join(", ", args.Select(x => x?.GetType().ToString() ?? "null").ToArray())})";
-            }
 
             // Check the cache if we already found a match for this hook
-            if (HooksCache.ContainsKey(fullName))
+            HookCache cache;
+            List<HookMethod> methods = HooksCache.GetHookMethod(name, args, out cache);
+            if (methods != null)
             {
-                return HooksCache[fullName];
+                return methods;
             }
-
-            List<HookMethod> methods;
             var matches = new List<HookMethod>();
-
             // Get all hook methods that could match, return an empty list if none match
             if (!Hooks.TryGetValue(name, out methods))
             {
@@ -377,8 +257,8 @@ namespace Oxide.Core.Plugins
                         }
                     }
 
-                    ArrayPool.Free(hookArgs);
-                }
+                        ArrayPool.Free(hookArgs);
+               }
                 else
                 {
                     hookArgs = args;
@@ -396,8 +276,6 @@ namespace Oxide.Core.Plugins
                     // Should we determine the level and call the closest overloaded match? Performance impact?
                     overloadedMatch = h;
                 }
-
-                
             }
 
             if (exactMatch != null)
@@ -412,7 +290,9 @@ namespace Oxide.Core.Plugins
                 }
             }
 
-            return HooksCache[fullName] = matches;
+            cache.SetupMethods(matches);
+
+            return matches;
         }
 
         protected virtual object InvokeMethod(HookMethod method, object[] args) => method.Method.Invoke(this, args);
